requirements.txt
streamlit>=1.28.0
faiss-cpu>=1.7.4
beautifulsoup4>=4.12.0
requests>=2.31.0
google-genai>=1.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
lxml>=4.9.0
